pyproject.toml
[tool.poetry]
name = "linebot"
version = "0.1.0"
description = ""
authors = ["Rick <rickyang2910@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
line-bot-sdk = "^3.9.0"
loguru = "^0.7.2"
python-decouple = "^3.8"
gspread = "^6.1.0"
oauth2client = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
